Feat: Quantization on byte (#39)
* POC code for int8 quantization


* remove unused methods in quantizer


* bug fix: negative values for int8


* resolve conflicts


* Resolve conflict


* explicilty make quantizer a utility class


* Bug fix: incorrect recall when using full 8-bit


* Refactor: refactor 3.0f as a constant


* Refactor: introduce static class for bit operation


* spotless apply


* Implement unsigned byte comparator


* method rename


* bug fix: sparse vector intialize


* Bug fix: import


* Bug fix: sort order


* Provide multiply function


* Bug fix: set get unsigned byte to be public


* bug fix: gradle assemble


---------

+++ src/main/java/org/opensearch/neuralsearch/sparse/common/DocFreqIterator.java
@@ -10,5 +10,5 @@ import org.apache.lucene.search.DocIdSetIterator;
  * DocFreqIterator is a DocIdSetIterator that also provides the frequency of the term in the document
  */
 public abstract class DocFreqIterator extends DocIdSetIterator {
-    public abstract float freq();
+    public abstract byte freq();
 }
